third_party/WebKit/LayoutTests/external/wpt/dom/nodes/Document-contentType/contentType/contenttype_javascripturi-expected.txt

This is a testharness.js-based test.
FAIL Javascript URI document.contentType === 'text/html' assert_equals: expected "text/html" but got ""
Harness: the test ran to completion.
